+++ src/components/TopNav/helpers/TopNavAvatar.tsx
